ThemeContext added to help in managing light and dark mode in application
